chore(connector): plugin-check compliance + no-store cache headers (v0.1.2)
PCP compliance pass (no behaviour change):

* SettingsPage: explicit (int) cast on $secondsLeft before sprintf —
  satisfies WordPress.Security.EscapeOutput.OutputNotEscaped.
* Signer: wrap the InvalidArgumentException message in sprintf() so
  WordPress.Security.EscapeOutput.ExceptionNotEscaped no longer flags
  the SODIUM_CRYPTO_SIGN_SECRETKEYBYTES constant concatenation.
* Plugin header License: "Proprietary" -> "GPL v2 or later" + add
  License URI, matching the WP plugin convention.
* New readme.txt with the wp.org headers (Stable tag, Tested up to,
  License) Plugin Check requires.

Cache-header hardening (real fix):

* RestRouter::applyNoCacheHeaders sends Cache-Control: no-store,
  no-cache, must-revalidate, max-age=0, private + Pragma + Expires
  on every successful defyn-connector/v1 response. Live evidence —
  smartcoding.com.au (2026-06-04, WP.com Atomic): Batcache served
  pre-handshake 404 responses on /status for the dashboard's signed
  sync calls for the full 5-minute TTL after a successful handshake.
  The fix is hosting-environment-agnostic — also protects against
  Kinsta full-page cache, Cloudflare edge, WP-Rocket, LiteSpeed,
  NGINX micro-cache, et al.

+++ packages/connector-plugin/src/Admin/SettingsPage.php
@@ -55,13 +55,13 @@ final class SettingsPage
         echo '<h2>' . esc_html__('Connection', 'defyn-connector') . '</h2>';
 
         if ($current === 'awaiting-handshake') {
-            $secondsLeft = max(0, $expires - time());
+            $secondsLeft = (int) max(0, $expires - time());
             echo '<p><strong>' . esc_html__('Connection code:', 'defyn-connector') . '</strong></p>';
             echo '<p style="font-size:2rem;font-family:monospace;">' . esc_html($code) . '</p>';
             echo '<p>' . sprintf(
                 /* translators: %d: seconds remaining until the code expires */
                 esc_html__('Expires in %d seconds. Paste this code into the DefynWP Dashboard "Add Site" form.', 'defyn-connector'),
-                $secondsLeft
+                (int) $secondsLeft
             ) . '</p>';
 
             self::renderResetForm();
